+++ packages/event-bus-local/src/index.ts
@@ -11,3 +11,4 @@ const moduleDefinition: ModuleExports = {
 }
 
 export default moduleDefinition
+export * from "./initialize"