+++ packages/medusa-plugin-algolia/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.2.10-rc.3
+
+### Patch Changes
+
+- Updated dependencies [[`a0c919a8d`](https://github.com/medusajs/medusa/commit/a0c919a8d01ca5edf62336de48e9a112e3822f38), [`bd12a9508`](https://github.com/medusajs/medusa/commit/bd12a95083b69a70b83ad38578c5a68738c41b2b)]:
+  - @medusajs/utils@0.0.2-rc.2
+  - @medusajs/modules-sdk@0.1.0-rc.4
+
 ## 0.2.10-rc.2
 
 ### Patch Changes